requirements.txt
aiofile
aiofiles
aiohttp>=2.2.5
alphabet_detector
beautifulsoup4
bleach
cachetools
configparser
currencyconverter
emoji
feedparser
future
git+https://github.com/python-telegram-bot/ptbcontrib.git
gitpython
gpytranslate
markdown2
multicolorcaptcha
pillow>=4.2.0
psutil
psycopg2-binary
pyYAML>=5.1.2
pyrate-limiter
python-telegram-bot==13.9
requests
spamwatch
speedtest-cli
sqlalchemy==1.4.11
telethon==1.24.0
ujson
pycryptodome
base58
py-tgcalls==0.9.1
pytimeparse
git+https://github.com/Itz-fork/py-trans.git
python-dotenv 
numpy
youtube-search-python
yt-dlp
